Update mAh counter to get from Mavlink

+++ telemetry.h
@@ -186,7 +186,7 @@ typedef struct {
 	uint32_t SP;
 	uint32_t SE;
 	uint32_t SH;
-        float total_amps;
+        int32_t total_amps;
 #endif
 
 #if defined(LTM)
